Picking up items
Once the player picks up an item it is removed from Block_list and therefore gets removed from the screen. Once the Backspace Key is pressed this block is then removed form the playerss carry_item_List and back into Block_List and is therefore drawn back onto the screen.